Fix: some log messages in the installation script

+++ lg/lg_install.sh
@@ -9,11 +9,15 @@ reset='\e[0m'
 LG_PATH=$(pwd)
 LG_MODULES=$LG_PATH/modules
 
+echo -e "${yellow}Please, provide your password to be able to install Lua Generate${reset}"
 sudo cp -r $LG_MODULES /usr/lib/lua/5.3/
 
 # it will make sure that the command will not be
 # appended multiple times to the file
-rm lg
+if [[ ! -f lg ]]; then
+  rm lg
+fi
+
 touch lg
 chmod +x ./lg
